+++ commands/tools/services.js
@@ -39,7 +39,7 @@ module.exports = class ServicesCommand extends commando.Command {
           style: { head: [], border: [] }
         })
         cpuTable.push(
-          [{ colSpan: 2, content: 'Sorted by CPU usage' }],
+          [{ colSpan: 2, content: 'Sorted by CPU usage', hAlign: 'center' }],
           ...data
             .sort((x, y) => x.cpu - y.cpu)
             .reverse()
@@ -51,7 +51,7 @@ module.exports = class ServicesCommand extends commando.Command {
           style: { head: [], border: [] }
         })
         memTable.push(
-          [{ colSpan: 2, content: 'Sorted by RAM usage' }],
+          [{ colSpan: 2, content: 'Sorted by RAM usage', hAlign: 'center' }],
           ...data
             .sort((x, y) => x.mem - y.mem)
             .reverse()